(feat) Increase termux dir on top of file
As than command name, now dir termux configuration is place into top of file to
be easy to reuse on rest of code, but, with advantage to validate error
path, seting empty content in this case.

Too, i increase quotes into command line name to keep everything with
correct values to be storaged on variavel.