(refactor/feat) Increase SHOW-ALL option to separate responsabilities
To be more easier to create a condicional and to be more intuitive to
use, i increased a option to replace [get -a], now called of [show-all]
to be more explicit as then possible. It will do the same thing that
another, but with less code and better to use.

On GET option i removed the condicional to accept it, now only with -c
option to show commented options.